+++ server/handle_default.go
@@ -13,10 +13,11 @@ type (
 		locator ComponentLocator
 	}
 	getDefaultHandler struct {
+		routeMap restful.RouteMap
 	}
 
 	Default struct {
-		Message string
+		Paths []string
 	}
 )
 
@@ -24,12 +25,18 @@ func newDefaultResource(loc ComponentLocator) *defaultResource {
 	return &defaultResource{locator: loc}
 }
 
-func (dr *defaultResource) Get(*restful.RouteMap, logging.LogSink, http.ResponseWriter, *http.Request, httprouter.Params) restful.Exchanger {
-	return &getDefaultHandler{}
+func (dr *defaultResource) Get(rm *restful.RouteMap, ls logging.LogSink, rw http.ResponseWriter, r *http.Request, p httprouter.Params) restful.Exchanger {
+	return &getDefaultHandler{
+		routeMap: *rm,
+	}
 }
 
 func (gdh *getDefaultHandler) Exchange() (interface{}, int) {
+	paths := make([]string, 0)
+	for _, re := range gdh.routeMap {
+		paths = append(paths, re.Path)
+	}
 	return Default{
-		Message: "all good",
+		Paths: paths,
 	}, 200
 }