+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)
 with respect to its command line interface and HTTP interface
 
+## [Unreleased](//github.com/opentable/sous/compare/0.5.112...master)
+### Changed
+* Client: 'sous artifact get' no longer requires -cluster flag.
+* Client: 'sous artifact get' now prints artifact information (digest, type).
+
 ## [0.5.112](//github.com/opentable/sous/compare/0.5.111...0.5.112)
 ### Changed
 * Client: when receiving bad content type print the HTTP status code and