+++ CHANGELOG.md
@@ -6,13 +6,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)
 with respect to its command line interface and HTTP interface
 
+## [0.5.114](//github.com/opentable/sous/compare/0.5.113...0.5.114)
+### Changed
+* Server: Default endpoint / implemented
+
+## [0.5.113](//github.com/opentable/sous/compare/0.5.112...0.5.113)
+### Changed
+* Both: Send the request URL back to the client and print to
+	console on deploy
+
 ## [0.5.112](//github.com/opentable/sous/compare/0.5.111...0.5.112)
 ### Changed
 * Client: when receiving bad content type print the HTTP status code and
   text along with that error message.
-* Both: Send the request URL back to the client and print to
-	console on deploy
-	
+
 ## [0.5.111](//github.com/opentable/sous/compare/0.5.110...0.5.111)
 ### Fixed
 * Client: builds using split container strategy were broken due to unique